+++ src/common/validation-tasks.js
@@ -12,7 +12,7 @@ function checkBooleanValue(boolValue, boolProp, resObj)
 	else
 	{
 		resObj.valid = false;
-		resObj.errorMessage = writeBooleanErrorText(boolProp);
+		resObj.errorMessage = initializeErrorText(boolProp, "must be True or False.");
 	}
 	
 	return checkRes;
@@ -73,24 +73,24 @@ function checkPercentageValue(numValue, numProp, resObj)
 	else
 	{
 		resObj.valid = false;
-		resObj.errorMessage = writePercentageErrorText(numProp);
+		resObj.errorMessage = initializeErrorText(numProp, "must be a valid percentage decimal between 0 and 1.");
 	}
 }
 
 
-function checkOffsetValue(numValue, numProp, resObj)
+function checkOffsetValue(numValue, numProp, rangeObj, resObj)
 {
 	var correctType = Number.isFinite(numValue);
 	var checkRes = false;
 	
-	if (correctType === true && numValue >= 0 && numValue <= 1000)
+	if (correctType === true && numValue >= rangeObj.minValue && numValue <= rangeObj.maxValue)
 	{
 		checkRes = true;
 	}
 	else
 	{
 		resObj.valid = false;
-		resObj.errorMessage = "OFFSET ERROR";
+		resObj.errorMessage = writeDecimalRangeErrorText(numProp, rangeObj);
 	}
 	
 	return checkRes;
@@ -133,42 +133,28 @@ function checkDayWeightsArrayValue(arrValue, arrProp, resObj)
 	else if (correctType === true)
 	{
 		resObj.valid = false;
-		resObj.errorMessage = writeDayWeightCountError(arrProp);
+		resObj.errorMessage = initializeErrorText(arrProp, "must have exactly 7 elements. One for each day of the week.");
 	}
 	else
 	{
 		resObj.valid = false;
-		resObj.errorMessage = writeArrayErrorText(arrProp);
+		resObj.errorMessage = initializeErrorText(arrProp, "must be a valid array object.");
 	}
 	
 	return checkRes;
 }
 
 
-function writeBooleanErrorText(vProp)
-{
-	var writeRes = initializeErrorText(vProp, "must be True or False.");
-	return writeRes;
-}
-
-
-function writeArrayErrorText(vProp)
-{
-	var writeRes = quoteProperty(vProp, "must be a valid array object.");
-	return writeRes;
-}
-
-
 function writeNumberErrorText(vProp)
 {
-	var writeRes = quoteProperty(vProp, "must be a positive whole number.");
+	var writeRes = initializeErrorText(vProp, "must be a positive whole number.");
 	return writeRes;
 }
 
 
 function writeNumberTooLargeErrorText(vProp)
 {
-	var writeRes = quoteProperty(vProp, "is too large to be used safely. Please use a lower number.");
+	var writeRes = initializeErrorText(vProp, "is too large to be used safely. Please use a lower number.");
 	return writeRes;
 }
 
@@ -177,7 +163,7 @@ function writeNumberRangeErrorText(vProp, vRange)
 {
 	var writeRes = "";
 	
-	writeRes += quoteProperty(vProp, "must be a positive, whole number between ");
+	writeRes += initializeErrorText(vProp, "must be a positive, whole number between ");
 	writeRes += appendRangeNumbers(vRange);
 	
 	return writeRes;
@@ -188,34 +174,19 @@ function writeDecimalRangeErrorText(vProp, vRange)
 {
 	var writeRes = "";
 	
-	writeRes += quoteProperty(vProp, "must be a decimal value between ");
+	writeRes += initializeErrorText(vProp, "must be a decimal value between ");
 	writeRes += appendRangeNumbers(vRange);
 	
 	return writeRes;
 }
 
-
-function writePercentageErrorText(vProp)
-{
-	var writeRes = quoteProperty(vProp, "must be a valid percentage decimal between 0 and 1.");
-	return writeRes;
-}
-
-
 function writeDateErrorText(vProp)
 {
 	var writeRes = "";
 	
-	writeRes += quoteProperty(vProp, "must be a valid date string. ");
+	writeRes += initializeErrorText(vProp, "must be a valid date string. ");
 	writeRes += "(eg. '2021-06-19')";
 	
-	return writeRes;
-}
-
-
-function writeDayWeightCountError(vProp)
-{
-	var writeRes = quoteProperty(vProp, "must have exactly 7 elements. One for each day of the week.");
 	return writeRes;
 }
 
@@ -232,3 +203,16 @@ function appendRangeNumbers(rNums)
 	var appendRes = [rNums.minValue, "and", rNums.maxValue].join(" ");
 	return appendRes;
 }
+
+
+
+module.exports =
+{
+	checkBoolean: checkBooleanValue,
+	checkNumber: checkNumberValue,
+	checkRange: checkRangeValue,
+	checkPercentage: checkPercentageValue,
+	checkOffset: checkOffsetValue,
+	checkDateString: checkDateStringValue,
+	checkDayWeightsArray: checkDayWeightsArrayValue
+};